Move optimize test out of context.rs
